requirements.txt
Django~=2.0.6
DjangoRestFramework~=3.9.0